Superstructure code cleanup (#400)
* Fold exchanges methods into dataframe.py

* Move df-to-array method into dataframe.py

* Remove unused methods from dataframe.py

* Clarify index error for header lookup

* Cleanup unused activities.py methods
