+++ activity_browser/app/bwutils/superstructure/__init__.py
@@ -1,9 +1,7 @@
 # -*- coding: utf-8 -*-
-from .activities import all_flows_found, all_activities_found, fill_df_keys_with_fields
 from .dataframe import (
-    build_superstructure, scenario_names_from_df, superstructure_from_arrays
+    scenario_names_from_df, superstructure_from_arrays
 )
 from .excel import import_from_excel, get_sheet_names
-from .exchanges import all_exchanges_found, filter_existing_exchanges
 from .mlca import SuperstructureMLCA, SuperstructureContributions
 from .utils import SUPERSTRUCTURE
