+++ activity_browser/app/bwutils/superstructure/activities.py
@@ -4,7 +4,7 @@ from bw2data.backends.peewee import ActivityDataset
 import numpy as np
 import pandas as pd
 
-from .utils import FROM_ALL, TO_ALL, EXCHANGE_KEYS
+from .utils import FROM_ALL, TO_ALL
 
 
 FROM_ACT = pd.Index([
@@ -161,29 +161,6 @@ def get_relevant_flows(df: pd.DataFrame, part: str = "from") -> dict:
     return flows
 
 
-def convert_fields_to_key(df: pd.DataFrame) -> pd.Series:
-    """Converts the process fields to its actual key by matching the database."""
-    assert all_activities_found(df), "Some processes could not be found in the database"
-    matches = get_relevant_activities(df)
-    combinations = df.iloc[:, 0:3].apply(tuple, axis=1)
-    keys = pd.Series([matches[x] for x in combinations], dtype="object")
-    return keys
-
-
-def convert_key_to_fields(df: pd.DataFrame) -> pd.DataFrame:
-    """Converts the process fields to its actual key by matching the database."""
-    keys = set(df.iloc[:, 5])
-    dbs, codes = zip(*keys)
-    query = (ActivityDataset
-             .select()
-             .where((ActivityDataset.database.in_(set(dbs))) &
-                    (ActivityDataset.code.in_(set(codes))))
-             .namedtuples())
-    key_data = dict(constuct_ad_data(x) for x in query.iterator())
-    subdf = pd.DataFrame([key_data[x] for x in df.iloc[:, 5]], columns=df.columns[0:5])
-    return subdf
-
-
 def match_fields_for_key(df: pd.DataFrame, matchbook: dict) -> pd.Series:
     def build_match(row):
         if row.iat[4] == bw.config.biosphere:
@@ -202,19 +179,3 @@ def fill_df_keys_with_fields(df: pd.DataFrame) -> pd.DataFrame:
     matches.update(get_relevant_activities(df, "to"))
     df["to key"] = match_fields_for_key(df.loc[:, TO_ALL], matches)
     return df
-
-
-def fill_out_df_with_keys(df: pd.DataFrame) -> pd.DataFrame:
-    """Will attempt to fill out the name, product, category, location and
-    database fields using the 'from' and 'to' keys.
-
-    Will raise an Exception if any key is missing in the DataFrame.
-    """
-    assert df.loc[:, EXCHANGE_KEYS].notna().all().all(), "All keys should be known before running this method."
-    from_df = convert_key_to_fields(df.loc[:, FROM_ALL])
-    df[from_df.columns] = from_df
-    to_df = convert_key_to_fields(df.loc[:, TO_ALL])
-    df[to_df.columns] = to_df
-    return df
-
-
